service/requirements-dev.txt
bandit
mypy
pylint
pytest-playwright
ruff
